backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.1",
    "finnhub-python>=2.4.25",
    "hatchling>=1.27.0",
    "pydantic-settings>=2.12.0",
    "rag",
    "shared",
    "uvicorn>=0.38.0",
]

[project.scripts]
backend = "backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
rag = { path = "../rag" }
shared = { path = "../shared", editable = true }

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]
